requirements/base.txt
Django==2.1
Fabric3==1.13.1.post1
Markdown==2.6.11
django-filter==2.0.0
djangorestframework==3.8.2
flake8==3.5.0
psycopg2==2.7.5
raven==6.9.0
